+++ datagrowth/processors/input/__init__.py
@@ -1,1 +1,2 @@
 from .extraction import ExtractProcessor
+from .iterators import content_iterator