+++ tests/processors/tests/__init__.py
@@ -1,2 +1,3 @@
 from .base import TestProcessorBase, TestDatagrowthProcessorDjangoConfig
 from .extraction import TestExtractProcessor
+from .iterators import TestContentIteratorWithSendIterator, TestContentIteratorWithSendSerieIterator